Small fixes for macos (clang)
