+++ Examples/CMakeLists.txt
@@ -1,3 +1,6 @@
+# Explicitly list source files in this shared library
+# When a new source file is added, it must be added to this list.
+
 # generate root dictionary
 ROOT_GENERATE_DICTIONARY(ExamplesDict
     # headers to include in ROOT dictionary
@@ -9,8 +12,10 @@ ROOT_GENERATE_DICTIONARY(ExamplesDict
     MODULE Examples
 )
 # create shared library with ROOT dict
-add_library(Examples SHARED ExamplesDict)
-
+add_library(Examples SHARED
+  CaloDistanceToTime.cc
+  ExamplesDict
+)
 target_include_directories(Examples PRIVATE ${PROJECT_SOURCE_DIR}/..)
 target_include_directories(Examples PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
