+++ Tests/ToyMC.hh
@@ -8,7 +8,7 @@
 #include "KinKal/MatEnv/MatDBInfo.hh"
 #include "KinKal/MatEnv/DetMaterial.hh"
 #include "KinKal/MatEnv/SimpleFileFinder.hh"
-#include "KinKal/Trajectory/CaloDistanceToTime.hh"
+#include "KinKal/Examples/CaloDistanceToTime.hh"
 #include "KinKal/Trajectory/Line.hh"
 #include "KinKal/Trajectory/ParticleTrajectory.hh"
 #include "KinKal/Trajectory/PiecewiseClosestApproach.hh"
@@ -227,10 +227,10 @@ namespace KKTest {
     shmaxMeas.SetZ(endpos.Z()+clen_);
     // set the measurement time to correspond to the light propagation from showermax_, smeared by the resolution
     double tmeas = tr_.Gaus(shmaxtime+(shmaxMeas.Z()-shmaxTrue.Z())/cprop_,scitsig_);
-    
+
     // create the ttraj for the light propagation
     VEC3 lvel(0.0,0.0,cprop_);
-    
+
     // put in manual values
     // std::shared_ptr calod2t = std::make_shared<CaloDistanceToTime>(1.985, 85.76, 27.47);
     //CaloDistanceToTime calod2t(tmeas, 85.76, 27.47);
